requirements.txt
telethon >= 1.37.0
pyyaml >= 6.0.2
sqlalchemy >= 2.0.36
psycopg2-binary >= 2.9.10